requirements.txt
msal==1.28.0
requests==2.32.3
python-dotenv==1.0.1
pyttsx3==2.90
google-generativeai==0.8.3
rich==13.7.1
typer==0.9.0
colorama==0.4.6
vosk==0.3.45
sounddevice==0.4.7
numpy>=1.21.0
webrtcvad==2.0.10
httpx==0.24.1
# Additional dependencies for enhanced voice features
# Note: PyAudio requires PortAudio headers on Windows
# Alternative: Use sounddevice (already included) for audio input
speechrecognition==3.10.0

# Enhanced Luca Dependencies
# Natural Language Understanding
rasa==3.6.15
transformers>=4.30.0
torch>=2.0.0

# Memory and Context Management
sqlite3  # Built-in with Python

# Enhanced TTS
coqui-tts>=0.19.0
elevenlabs>=0.2.0

# Multimodal Features
google-calendar-api>=1.0.0
twilio>=8.0.0
pyautogui>=0.9.54

# Security and Privacy
keyring>=23.0.0
cryptography>=41.0.0

# Additional AI/ML
openai>=1.0.0
langchain>=0.1.0

# Advanced Features Dependencies
# Multimodal Awareness
opencv-python>=4.8.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0

# Voice Authentication
librosa>=0.10.0
scikit-learn>=1.3.0

# Task Automation
pyperclip>=1.8.2

# Gamification (built-in modules)
# json, random, time, threading - all built-in